Vercel error resolved

+++ src/app/product/api/[productId]/route.ts
@@ -1,3 +1,4 @@
+import { getCurrentUser } from '@/lib/firebase/admin';
 import prisma from '@/lib/prisma/prisma'
 import { NextRequest } from 'next/server'
 
@@ -19,12 +20,19 @@ export async function DELETE(
   _request: NextRequest,
   { params }: { params: { productId: string } }
 ) {
+  const currentUser = await getCurrentUser()
   const { productId } = params;
-  if (productId) {
-    await prisma.product.delete({
-      where: { id: Number(productId) },
-    })
-    return Response.json({ success: true, data: 'Product deleted!' }, { status: 202 })
+
+  if (productId && currentUser) {  
+    const deletedProduct = await prisma.product.delete({
+      where: { id: Number(productId), userId: currentUser.uid },
+    });
+
+    if (deletedProduct) {
+      return Response.json({ success: true, data: 'Product deleted!' }, { status: 202 })
+    }
+
+    return Response.json({ success: false, data: 'Product not found' }, { status: 404 })
   }
 
   return Response.json({ success: false, data: 'Product not deleted'}, { status: 400 })
